several bugs fixed, especially check if signal is below upper threshold (was using calibration instead of signal strength)

+++ scripts/scan_CMIP5.ksh
@@ -1,51 +1,19 @@
 #!/bin/ksh
 
-## test 1
-#varlist="intpp"
-#monlist="4" # 1 4 7 10"
-#modellist="NorESM1-ME"
-#sigmaS=500
-#sigmaT=10
-#realm=ocean
-
-#### test 2
-#varlist="tas"
-#monlist="3"
-#modellist="MPI-ESM-LR"
-#sigmaS=200
-#sigmaT=10
-#realm=atmosphere
-
-#### test 3
-#varlist="intpp"
-#monlist="7"
-#modellist="IPSL-CM5A-LR"
-#sigmaS=200
-#sigmaT=10
-#realm=ocean
-
-#### test 4
-#varlist="mrso"
-#monlist="7"
-#modellist="CCSM4"
-#sigmaS=200
-#sigmaT=10
-#realm=land
-
-
-
 ## scan automatically
 
 ### all models I ever found so far
 modellist="ACCESS1-0 ACCESS1-3 bcc-csm1-1 bcc-csm1-1-m BNU-ESM CanAM4 CanCM4 CanESM2 CCSM4 CESM1-BGC CESM1-CAM5 CESM1-CAM5-1-FV2 CESM1-FASTCHEM CESM1-WACCM CFSv2-2011 CMCC-CESM CMCC-CM CMCC-CMS CNRM-CM5 CNRM-CM5-2 CSIRO-Mk3-6-0 CSIRO-Mk3L-1-2 EC-EARTH FGOALS-g2 FGOALS-gl FGOALS-s2 FIO-ESM GEOS-5 GFDL CM2p1 GFDL-CM3 GFDL-ESM2G GFDL-ESM2M GFDL-HIRAM-C180 GFDL-HIRAM-C360 GISS-E2-H GISS-E2-H-CC GISS-E2-R GISS-E2-R-CC HadCM3 HadGEM2-A HadGEM2-AO HadGEM2-CC HadGEM2-ES inmcm4 IPSL-CM5A-LR IPSL-CM5A-MR IPSL-CM5B-LR MIROC4h MIROC5 MIROC-ESM MIROC-ESM-CHEM MPI-ESM-LR MPI-ESM-MR MPI-ESM-P MRI-AGCM3-2H MRI AGCM3-2S MRI-CGCM3 MRI-ESM1 NICAM-09 NorESM1-M NorESM1-ME"
 
 
-# selection for testing
-modellist="ACCESS1-0 GFDL-CM3 MIROC-ESM HadGEM2-ES-CC MPI-ESM-MR EC-EARTH IPSL-CM5A-LR"
 
-sigmaS=100
-sigmaT=10
+## selection for testing
+#modellist="ACCESS1-0 GFDL-CM3 MIROC-ESM HadGEM2-ES-CC MPI-ESM-MR EC-EARTH IPSL-CM5A-LR"
 
+
+
+
+# variables
 
 realm=atmosphere
 
@@ -54,18 +22,16 @@ varlist="  snw   lai  gpp npp rlus huss  prsn prw ps rlut rsut      snc snw   mr
 monlist="1 4 7 10"
 
 
-
-
-
 ## slow land and atm vars,
 #varlist="cSoil cVeg baresoilFrac grassFrac treeFrac shrubFrac vegFrac treeFracPrimDec treeFracPrimEver treeFracPrimSecDec treeFracPrimSecEver burntArea c3PftFrac c4PftFrac"
-#monlist="4"
+#monlist="13"
 
 
 
 # fast ocean vars
 
 # slow ocean vars
+
 
 
 
@@ -88,12 +54,9 @@ logfile=1
 
 
 
-### error check:
-modellist="ACCESS1-0 ACCESS1-3"
-varlist=snw
-monlist=10
-calc_new=1
-logfile=0
+sigmaS=100
+sigmaT=10
+
 
 
 
@@ -102,6 +65,8 @@ scen=rcp85
 threshcase=2
 
 outpath=/media/sf_D_DRIVE/datamining/edges/CMIP5scan
+hyperccpath=/home/sebastian/Abrupt/hypercc/bin
+
 
 
 thresh1=pi-control-max
@@ -116,16 +81,15 @@ fi
 
 
 
-
 for var in ${varlist}; do
 for model in ${modellist}; do
 for mon in ${monlist}; do
 
 echo ""
-echo "${var} ${model} ${scen} mon ${mon}"
+echo "${var} ${model} ${scen} mon ${mon}, scales: ${sigmaT} year ${sigmaS} km, thresholds: ${thresh1} ${thresh2}"
 
 
-# remove cache if too large
+# remove cache if too large (can also switch off cache in workflow.py)
 if [[ -f hypercc-cache.hdf5 ]]; then
   filesize=`du -k "hypercc-cache.hdf5" | cut -f1`
   if [[ ${filesize} -gt 100000 ]]; then
@@ -139,6 +103,15 @@ fi
 if [[ ! -f ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt || ${calc_new} == 1 ]]; then
 
 
+## clear figures (needs to have enough time for new time stamp)
+mkdir -p ${outpath}/figs/${var}
+rm -f ${outpath}/figs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}.png
+for type in signal years event_count regions peakiness maxTgrad timeseries; do
+  rm -f ${outpath}/singlefigs/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_${type}.png 
+done
+
+
+
 ## ocean (regridded)
 if [[ ${realm} == ocean ]]; then
 
@@ -147,22 +120,49 @@ if [[ ${realm} == ocean ]]; then
   rcppath=/media/sf_D_DRIVE/CMIP5/data_duplicate
   piCpath=${rcppath}
 
-
   files_rcp=`ls ${rcppath}/${var}_*mon_${model}_${scen}_${rea}_??????-??????.${extension} 2>/dev/null | wc -w `
   files_piC=`ls ${piCpath}/${var}_*mon_${model}_piControl_${rea}_??????-??????.${extension} 2>/dev/null | wc -w`
 
   if [[ ${files_rcp} -gt 0 && ${files_piC} -gt 0 ]]; then
     if [[ ${logfile} == 1 ]]; then
       mkdir -p ${outpath}/logs/${var}
-      ./bin/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} --extension ${extension} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} --lower-threshold ${thresh2}  > ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+        echo "${var} ${model} ${scen} mon ${mon}, scales: ${sigmaT} year ${sigmaS} km, thresholds: ${thresh1} ${thresh2}" > ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      if [[ ${mon} == 13 ]]; then
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --extension ${extension} --annual --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km \
+          --upper-threshold ${thresh1} --lower-threshold ${thresh2}  \
+          >> ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      else
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --extension ${extension} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km \
+          --upper-threshold ${thresh1} --lower-threshold ${thresh2}  \
+          >> ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      fi #annual
     else
-      ./bin/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} --extension ${extension} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} --lower-threshold ${thresh2} 
+      if [[ ${mon} == 13 ]]; then
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --extension ${extension} --annual --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km \
+          --upper-threshold ${thresh1} --lower-threshold ${thresh2} 
+      else
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath} \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --extension ${extension} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km \
+          --upper-threshold ${thresh1} --lower-threshold ${thresh2} 
+      fi #annual
     fi
+  else
+    echo "no files found"
   fi
   ## optional:  --output-folder ${outpath}
 
 
-  # land and atmo (not regridded)
+
+
+
+  ### land and atmo (not regridded)
 elif [[ ${realm} == land || ${realm} == atmosphere ]]; then
 
   rcppath=/media/sf_W_DRIVE/PROJECTS/CMIP5data/modeldata/${model}/${scen}/${var}
@@ -174,42 +174,73 @@ elif [[ ${realm} == land || ${realm} == atmosphere ]]; then
 #  piCpath=${rcppath}
 
 
-
   files_rcp=`ls ${rcppath}/${var}_*mon_${model}_${scen}_${rea}_??????-??????.nc 2>/dev/null | wc -w`
   files_piC=`ls ${piCpath}/${var}_*mon_${model}_piControl_${rea}_??????-??????.nc 2>/dev/null | wc -w`
   if [[ ${files_rcp} -gt 0 && ${files_piC} -gt 0 ]]; then
+
     if [[ ${logfile} == 1 ]]; then
-      mkdir -p ${outpath}/logs/${var}/
-      ./bin/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} --lower-threshold ${thresh2} --sobel-scale 1 km/year > ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
-    else
-      ./bin/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath}  report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} --lower-threshold ${thresh2} --sobel-scale 1 km/year 
-    fi
-  fi
+      mkdir -p ${outpath}/logs/${var}
+      echo "${var} ${model} ${scen} mon ${mon}, scales: ${sigmaT} year ${sigmaS} km, thresholds: ${thresh1} ${thresh2}" > ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      if [[ ${mon} == 13 ]]; then
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath} \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --annual --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} \
+          --lower-threshold ${thresh2} --sobel-scale 1 km/year \
+          >> ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      else
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath} \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} \
+          --lower-threshold ${thresh2} --sobel-scale 1 km/year \
+          >> ${outpath}/logs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_log.txt
+      fi #annual
 
 
+    else # no logfile
+      if [[ ${mon} == 13 ]]; then
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath} \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --annual --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} \
+          --lower-threshold ${thresh2} --sobel-scale 1 km/year 
+      else
+        ${hyperccpath}/hypercc --data-folder ${rcppath} --pi-control-folder ${piCpath} \
+          report --variable ${var} --model ${model} --scenario ${scen} --realization ${rea} \
+          --month ${mon} --sigma-t ${sigmaT} year --sigma-x ${sigmaS} km --upper-threshold ${thresh1} \
+          --lower-threshold ${thresh2} --sobel-scale 1 km/year 
+      fi #annual
 
-fi #calculate new
+    fi #logfile
 
+  else
+    echo "no files found"
+  fi #files exist
+
+
+fi #realm
 
 
 
 ## select figures and put in 1 figure
-##rm -f ${outpath}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}.png
+
 if [[ ${files_rcp} -gt 0 && ${files_piC} -gt 0 ]]; then
   convert \( signal.png years.png regions.png -append \) \
-  \( timeseries.png maxTgrad.png peakiness.png -append \) +append ${outpath}/figs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}.png
-
+  \( timeseries.png maxTgrad.png peakiness.png -append \) \
+  +append ${outpath}/figs/${var}/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}.png
 
   for type in signal years event_count regions peakiness maxTgrad timeseries; do
-    mv ${type}.png ${outpath}/singlefigs/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_${type}.png 
+    if [[ -f ${type}.png ]]; then
+      mv ${type}.png ${outpath}/singlefigs/${var}_${model}_${scen}_${rea}_mon${mon}_sigmaT${sigmaT}_sigmaS${sigmaS}_lowthreshcase${threshcase}_${type}.png 
+    fi
   done
 
-fi
+fi #rearrange plots
 
 
 
+else
+  echo "case done already"
 
-fi # plot new
+fi # calc new
 
 done #mon
 done #model
@@ -217,3 +248,69 @@ done #var
 
 
 exit
+
+
+
+# test 1
+#varlist="intpp"
+#monlist="4" # 1 4 7 10"
+#modellist="NorESM1-ME"
+#sigmaS=500
+#sigmaT=10
+#realm=ocean
+
+#### test 2
+varlist="tas"
+monlist="13"
+modellist="MPI-ESM-LR"
+sigmaS=200
+sigmaT=10
+realm=atmosphere
+calc_new=1
+logfile=0
+
+#### test 3
+#varlist="intpp"
+#monlist="7"
+#modellist="IPSL-CM5A-LR"
+#sigmaS=200
+#sigmaT=10
+#realm=ocean
+
+#### test 4
+#varlist="mrso"
+#monlist="7"
+#modellist="CCSM4"
+#sigmaS=200
+#sigmaT=10
+#realm=land
+
+varlist="tas"
+monlist="1"
+modellist="ACCESS1-3" 
+sigmaS=100
+sigmaT=10
+realm=atmosphere
+calc_new=1
+logfile=0
+
+### testing:
+varlist="clwvi"
+monlist="10"
+modellist="CCSM4"
+sigmaS=100
+sigmaT=10
+realm=atmosphere
+calc_new=1
+logfile=0
+
+
+### testing:
+varlist="rsuscs"
+monlist="7"
+modellist="IPSL-CM5A-LR"
+sigmaS=100
+sigmaT=10
+realm=atmosphere
+calc_new=1
+logfile=0
